Move .h files from include/ to include/pathie/
