+++ include/pathie/entry_iterator.hpp
