+++ include/pathie/errors.hpp
