+++ include/pathie/pathie_ofstream.hpp
