+++ src/temp.cpp
@@ -1,4 +1,4 @@
-#include "../include/temp.hpp"
+#include "../include/pathie/temp.hpp"
 #include <sstream>
 #include <cstdlib>
 #include <cstring>
