+++ test/pathie_streams.cpp
@@ -30,8 +30,8 @@
 #include <cstdlib>
 #include <cstring>
 #include <locale>
-#include "../include/pathie_ifstream.hpp"
-#include "../include/pathie_ofstream.hpp"
+#include "../include/pathie/pathie_ifstream.hpp"
+#include "../include/pathie/pathie_ofstream.hpp"
 #include "testhelpers.hpp"
 
 #if defined(_PATHIE_UNIX)
